movies_order/test_api/requirements.txt
django==1.8.6
djangorestframework==3.3.2
requests==2.9.1
beautifulsoup4==4.5.1
mysql-python==1.2.5
Pillow==3.3.0
celery==3.1.23
redis==2.10.5
